requirements.txt
Flask==2.3.3 
mysql-connector-python==8.0.33 
bcrypt==4.0.1 
scikit-learn==1.3.0 
numpy==1.25.2
python-dotenv==1.0.0